+++ variables.tf
@@ -87,6 +87,12 @@ variable "db_admin_username" {
   default     = "postgres"
 }
 
+variable "db_name" {
+  description = "The name of the Orca database within the RDS cluster. Default set to PREFIX-disaster_recovery in main.tf."
+  type        = string
+  default     = null
+}
+
 variable "default_multipart_chunksize_mb" {
   type        = number
   description = "The default maximum size of chunks to use when copying. Can be overridden by collection config."