+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and includes an additional section for migration notes.
 - *ORCA-257* orca_catalog_reporting lambda now returns data from actual catalog.
 
 ### Changed
+- *ORCA-297* Default database name is now PREFIX-disaster_recovery
 - *ORCA-287* Updated copy_to_glacier and extract_filepaths_for_granule to [new Cumulus file format](https://github.com/nasa/cumulus/blob/master/packages/schemas/files.schema.json). 
 
 ### Migration Notes
@@ -41,6 +42,7 @@ and includes an additional section for migration notes.
 - These are the new variables added:
   - db_admin_username (defaults to "postgres")
   - db_host_endpoint (Requires a value. Set in terraform.tfvars to your RDS Cluster's endpoint, similar to "PREFIX-cumulus-db.cluster-000000000000.us-west-2.rds.amazonaws.com")
+  - db_name (Defaults to PREFIX-disaster_recovery. If preserving a database from a previous version of Orca, set to disaster_recovery.)
   - api_gateway_policy_vpc_id
 - Add the following ORCA required variables definition to your `variables.tf` or `orca_variables.tf` file.